BigDecimal

+++ BDTrabPt3/src/model/Funcionario.java
@@ -5,6 +5,8 @@
  */
 package model;
 
+import java.math.BigDecimal;
+
 /*CREATE TABLE FUNCIONARIO (
     CPF CHAR(14) NOT NULL,
     NOME VARCHAR2(40) NOT NULL,
@@ -28,12 +30,12 @@ public class Funcionario {
     private String cpf;
     private String nome;
     private String email;
-    private int telefone1;
-    private int telefone2;
+    private BigDecimal telefone1;
+    private BigDecimal telefone2;
     private String funcao;
     private String funcionarioGerenteCpf;
 
-    public Funcionario(String cpf, String nome, String email, int telefone1, int telefone2, String funcao, String funcionarioGerenteCpf) {
+    public Funcionario(String cpf, String nome, String email, BigDecimal telefone1, BigDecimal telefone2, String funcao, String funcionarioGerenteCpf) {
         this.cpf = cpf;
         this.nome = nome;
         this.email = email;
@@ -72,19 +74,19 @@ public class Funcionario {
         this.email = email;
     }
 
-    public int getTelefone1() {
+    public BigDecimal getTelefone1() {
         return telefone1;
     }
 
-    public void setTelefone1(int telefone1) {
+    public void setTelefone1(BigDecimal telefone1) {
         this.telefone1 = telefone1;
     }
 
-    public int getTelefone2() {
+    public BigDecimal getTelefone2() {
         return telefone2;
     }
 
-    public void setTelefone2(int telefone2) {
+    public void setTelefone2(BigDecimal telefone2) {
         this.telefone2 = telefone2;
     }
